Automatically backed up by Learn

+++ db/migrate/001_create_costumes.rb
@@ -10,8 +10,8 @@
 class CreateCostumes < ActiveRecord::Migration[5.1]
   def change
     create_table :costumes do |t|
-      t.string :name,
-      t.decimal :price,
+      t.string :name
+      t.float :price
       t.string :image_url
     end
   end
